Move Source, Target and ValueConverter mocks into own classes
